Modified build.xml to use cactify to reintegrate test classes.  Conditional tomcat4 or tomcat5 container.  Added jetty jar to test libs so lifeblog tests will work in cactus.  Should look into this further as this may cause jetty to be loaded inside tomcat during tests which is probably not too good.
BlogPlugin currently allows for setting blog title and pinging Technorati and Weblogs.com based on a BlogPreferences object.  Progress on lifeblog integration, user blog list based on existence of a BlogPreferences object in the main document of the blog web, for instance Blog.WebHome.  Some shuffling of integration tests.

Adding .classpath and .project files which seem to be for eclipse.

git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki/trunk@490 f329d543-caf0-0310-9063-dda96c69346f