- Deprecated the current constructors - Added 2 new constructors:
public XWiki(XWikiConfig config, XWikiContext context) 
public XWiki(XWikiConfig config, XWikiContext context,
    XWikiEngineContext engine_context, boolean noupdate)

Thus the "user" must start by creating a XWikiConfig, then a XWikiContext, possibly an EngineContext and then only he can create a XWiki object. The idea is that the creation of a config should be left to the config object and NOT to the Wiki class (it's not its role).

This leads to a better design. In addition to being more flexible this will allow for easier unit testing as we will not need a xwiki.cfg file anymore (it will still be used for functional testing though).


git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki/trunk@582 f329d543-caf0-0310-9063-dda96c69346f