XWIKI-1317: Move Hibernate mapping files out of the XWiki JAR and into the XWiki WAR
* Reverted commit:

"I think I jumped the gun on this. While moving all config files together is nice, this move makes it harder to reuse the xwiki JAR. For example in our M2 build we're using the XWiki JAR to load XML page. With the move this requires to have the hibernate mapping files in order to perform the loading. I think we should move the config files when we create the storage component.

For now, users who want to override the mapping files can do so by copying them in WEB-INF/classes. Files in there take precedence over those in the xwiki JAR."

git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki/trunk@3566 f329d543-caf0-0310-9063-dda96c69346f