XWIKI-1885: Sometimes some Radeox macros fail to render and Radeox filter exceptions appear in the XWiki logs
My take is that the errors are happening when there are several threads calling XWikiRadeoxRenderEngine at the same time (note that a new instance of XWikiRadeoxRenderEngine is created at each render call) . When this happens the init() method is called. There's a test to check if the filter pipe is initialized or not but this test isn't thread safe so what happens is that one thread will start reinitializing the filter pipe. Then the second thread sees a non null filter pipe and start the rendering. The problem is that the filters have not finished initializing in the first thread and thus the rendering fails randomly...


git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki-platform/core/trunk@6057 f329d543-caf0-0310-9063-dda96c69346f