XWIKI-1885: Sometimes some Radeox macros fail to render and Radeox filter exceptions appear in the XWiki logs
I hope I really fixed it now without adding a new threading issue. I tested it with JMeter with 10 concurrent threads and I didn't get any exception in the logs. It would have been better to only have a single XWikiRadeoxRenderEngine instance (instead of one per render() call!) but that object doesn't look thread safe (I think it simply has to do with the XWikiContext being stored in it, if it could be just passed in method calls without storing it it would become thread safe).


git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki-platform/core/trunk/xwiki-core@6167 f329d543-caf0-0310-9063-dda96c69346f