XWIKI-1645: build time/test jars should not be packaged with xwiki + possible bloat
Reverted the move to provided scope for the portlet api jar. It's not currently possible since XWiki Core has code that uses the Portlet API and this code is executed even when run in a Servlet container. Removing this code will need an important rewrite that can only be done with the new XWiki 2.0 architecture.

git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki-platform/core/trunk@6818 f329d543-caf0-0310-9063-dda96c69346f