XTJETTY-6: Increase the default maximum form content size on Jetty * Fixed regression introduced in previous commit * Fixed invalid comment in start_xwiki_debug.bat * Added explanation as to why we introduced that new Jetty parameter
Note1: We need to refactor the debug scripts so that they use the main scripts to prevent duplication
Note2: We need to add explanation why we use the LANG=fr_FR.ISO8859-1 param in the start_xwiki.sh script. I would add it if I knew why we're doing this.

git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki-platform/xwiki-tools/trunk@8336 f329d543-caf0-0310-9063-dda96c69346f