Fixing the current Container implementation for Servlets using ThreadLocal since right now it has no chance of working...
Note 1: There's a vote going on on the list to decide if we want to keep this model of not having to pass the Container object to all methods or not.

Note 2: I haven't tested the new ThreadLocal implementation yet. I'll test it when it's used for the first time. Right now it should be viewed as a POC only.

Note 3: I still need to find a way to get the correct implementation of the Container API into the component that requires it. Right now the componentn needs to specify a hint (like "servlet") but that's not the right way of doing it since the component should be agnostic WRT the implementation passed to it.

git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki-platform/core/trunk@9133 f329d543-caf0-0310-9063-dda96c69346f