XWIKI-2356: Refactoring of the XMLRPC subsystem in order to handle proper setup/cleanup of the XWiki context Patch from Fabio Mancinelli, lightly reviewed by Vincent Massol
I'm not too comfortable with the hacks we've had to do to allow intercepting the requests since those hacks have a good chance of breaking when we upgrade Axis in the future and it doesn't look like Axis was done to be extended this way from the overly complex code I could see.

That said I'm not familiar enough with Axis to suggest any alternative and I believe Fabio did his best to make it clean and couldn't find a better solution so I'm ok to commit this for now. Fabio asked on the Axis mailing list about the best way to achieve this so we need to follow up on this.

Note: I need this refactoring to introduce the new ThreadLocal container objects in the XMLRPC servlet threads.


git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki-platform/core/trunk/xwiki-core@9839 f329d543-caf0-0310-9063-dda96c69346f