XWIKI-3028: XMLRPC Subsystem refactoring/improvement XWIKI-3029: XMLRPC getObject() does not return any information about DBList, DBTreeList, etc. values.
(XWIKI-3028) Definition of the XmlRpcApi.

(XWIKI-3029) The XWikiObject class has been extended with a mechanism for retrieving metadata associated to properties. This is necessary for attaching to a property which is a DBList, for example, the list of the allowed values. Meta-properties are stored as normal object properties but they have a special suffix. The getProperties method does not return these meta-properties. In order to access to meta-properties additional methods are provided.

git-svn-id: https://svn.xwiki.org/svnroot/xwiki/platform/core/trunk@14895 f329d543-caf0-0310-9063-dda96c69346f