XWIKI-3360 Including a document with first or second heading level breaks the including document section edit links (Syntax 1.0)
Internal count of section was incremented for headings of included documents while they are not part of the document and thus not subject to section edit in an included context. In doubt I left the comment related to the increment, though I am not sure it is relevant.

Merged from branch 1.8


git-svn-id: https://svn.xwiki.org/svnroot/xwiki/platform/core/trunk@17630 f329d543-caf0-0310-9063-dda96c69346f