XABLOG-51: Create new post form triggers a JS error in IE6/IE7 Fixed.
The onsubmit listener was supposed to save a HTTP request + redirect, since JavaScript can compute the correct target location.
However, this means duplicating the velocity code in CreatePost, and any code duplication is a source of bugs, as this issue proves.
I removed this behavior since an extra request is better than code duplication (especially in a different programming language).

Slight performance improvement by submitting the form to xpage=plain, which does not render all the supporting UI.

git-svn-id: https://svn.xwiki.org/svnroot/xwiki/platform/xwiki-applications/trunk/blog@21972 f329d543-caf0-0310-9063-dda96c69346f