XWIKI-4325: "Failed to find parser for syntax []. Defaulting to xwiki/2.0 syntax." error
Note: I haven't found a generic way to solve the problem since an empty string value could be used to override a value defined in xwiki.properties for example. We would need the notion of initialized and nont-initialized properties in the wiki to solve this cleanly but it's more work and we need this fix for 2.0RC2.

git-svn-id: https://svn.xwiki.org/svnroot/xwiki/platform/core/trunk@23486 f329d543-caf0-0310-9063-dda96c69346f