XWIKI-4351: XWiki.WikiMacroClass's macro code field should be presented in a larger text area. XWIKI-4466: Lost nice display in view mode of boolean values from wiki macro and wiki macro parameter classes.
* Fixed.

NOTE: These issues were assumed to be fixed a long ago. But since xwiki packager plugin is dependent on xwiki-core 2.0, XE ends up with earlier versions of WikiMacroClass and WikiMacroParameterClass class definitions (when XE is bundled). Theoretically this should not have been a problem since when XE starts up it should update these classes. But due to a problem with the behaviour of BaseClass's add***Field() methods, these classes were not updated correctly.

git-svn-id: https://svn.xwiki.org/svnroot/xwiki/platform/xwiki-tools/trunk/xwiki-packager-plugin@24753 f329d543-caf0-0310-9063-dda96c69346f