XWIKI-5467: Attachments uploaded through REST API are saved twice
The previously used document saving code generates two versions of the uploaded attachment.
Changed the document saving code in order to reflect the current status of the code in com.xpn.xwiki.webUploadAction.uploadAttachment() which was taken as a blueprint for attachment uploading logic.

This is an additional commit that fixes the problem also in AttachmentsResource.addAttachment() and factorize the common code in a method storeAttachment() that will be used in both AttachmentsResource.addAttachment() and AttachmentResource.putAttachment()

git-svn-id: https://svn.xwiki.org/svnroot/xwiki/platform/core/trunk@30874 f329d543-caf0-0310-9063-dda96c69346f