XWIKI-5786 Deprecate Document(Save|Update|Delete)Event and add new Document(Created|Updated|Deleted)Event in xwiki-bridge instead
Fix cyclic dependency I've introduced in rev 33349.

Deprecating AbstractDocumentEvent that was in xwiki-observation-api (moved in xwiki-legacy)  and introducing a new one in xwiki-model. It means that code that extended the old AbstractDocumentEvent will need to be updated to use the new one (will have to document it in the release notes).



git-svn-id: https://svn.xwiki.org/svnroot/xwiki/platform/core/trunk@33356 f329d543-caf0-0310-9063-dda96c69346f