Keep testing the inline action (along with the edit action) in the Blog application for backward compatibility.
Removed "display" property from SheetClass. All sheets should be displayed in-line (as documents). Velocity templates are responsible for the page layout. In the future I'll add the posibilty for a sheet to specify/overwrite the velocity template to use (e.g. you create a view sheet but you don't want to use view.vm but a different template/layout).

Revert most of the changes I made to the velocity templates. At this point only edit.vm is aware of the sheets, because it needs to include the editinline.vm when there is an edit sheet available.