XWIKI-14763: Use the document's locale to find the title (#614)
* Use the document's locale to find the title
The search result's title field is empty because it's multi-lingual field.
This causes the title in solr search results to be null. So use the locale
field to look up the correct title field to populate the search result with.
* Use getFirstValue method instead of get
* Go back to the title alias field in qf
The title field at query time is an alias for all of the localized versions,
this aliasing is done by the XWikiDismaxQParserPlugin. So we don't need to look
at the title_ version in qf expression.