XWIKI-6665: Saving edits after lock expiration doesn't warn the user that the page may be stale (#1075)
## Issue

https://jira.xwiki.org/browse/XWIKI-6665

## Design of the solution

The proposed solution is to record in a field of the editor the last version of the working document and to send it to the server when saving.
The server will compare this version number against the latest version of the document: if the latest version is different than the one sent on the request it means someone certainly save the document in the meantime. We also compare the edition dates: in some edge cases we could have the same revision numbers that are not the same revisions (e.g. a deleted and recreated document).
We then compute a content and an object diff to remove the false positive: the doc could have been saved in the meantime because of an attachment added.
Finally in case of conflict we send a 409 status with a JSON containing both the previous and the latest version number.
In case of success we send a JSON the latest version number, so that we can update it.

## Changes

  * XWikiDocument.java: add some utility method to be able to compute the next version, without updating it. Handle of errors in getContentDiff.
  * XWikiAction.java: add an utility method to send a JSON answer.
  * SaveAction.java: add the new method to compute the conflict and change the answer in case of success.
  * SaveAndContinueAction.java: manage the errors in case of conflict.
  * xwiki-meta.js: Allow to bump and set a new version. Fire an event when it's the case. New attribution: isNew to know if a page is already created or not.
  * dataeditor.js: Bump the version when an object is added or removed, since it saves the document. 
  * actionButton.js: 
    * Get the current document version and listen on the new event to update it when it's needed.
    * Create a new form input with this version, the date of the edition and the edit status to provide it when saving.
    * Handle the 409 status answer by displaying a popup with the conflict message and by disabling the form/editor.
    * Save and view now also send an AJAX request to save, instead of submitting the form like it did.
    * Handle the various modes "Save and view / Save and continue / preview". 

## Screenshot

![warning_conflict](https://user-images.githubusercontent.com/1478232/54133573-9ddd3500-4416-11e9-8de5-ee10fdaf7d72.png)

## Tests

Some selenium tests have been executed to ensure the changes: namely EditWikiTest, InlineEditTest, ObjectEditTest, CompareVersionsTest, LanguageTest.