XWIKI-15450: Browser cache for attachments is never invalidated (needs browser clear cache) (#1092)
Fallback on `createAttachmentRevisionURL` when creating an attachment URL for download. This is an easy solution which use the already existing mechanism to add the revision number at the end of the URL. It should both prevents the cache mechanism in case of a new version has been updated, and use the cache mechanism when downloading several times the same attachment.