XWIKI-16447: Notifications related errors when upgrading XWiki from 8.4.6 to 10.11.8.
Using Collections.emptyList() is dangerous when some code expect to be able to call add() on the collection afterwards. The alternative is to recreate a new array list each time we add a new element into the list in the fear we might have a "read-only" list. This would be very uncomfortable.

(cherry picked from commit c6765f0194476a6127ba21c97cf948cfce0bf953)