XWIKI-17038: The votes are lost after renaming the page
Implements XObjectAverageRatingManager#moveAverageRatings and update the AverageRatings XObject entityReference field to reflect the new reference of the moved entities (the move Page as well as sub-entities).
As a consequence, empty constructor added to  UpdateAverageRatingFailedEvent and UpdatedAverageRatingEvent, and the data accompanying the average ratings event is now a list of AverageRatings to reflect the fact that multiple ratings can be modified on a single "transaction".