XWIKI-17975: Office export is using way too much memory
This work consists mainly in refactoring OfficeConverter API to not
return a byte[] that has to be kept in memory, but to return instead a
proper OfficeConverterResult object that contains the paths of the
various files resulting from the conversion. This new API is then used
wherever it's needed and the cleanup of the temporary files is only
performed at the very end of the usecases.

More specifically this has been done:
  * Add class OfficeConverterResult
  * Use it as return type of OfficeConverter#convertDocument and
    deprecate OfficeConverter#convert
  * Refactor OfficeDocument interface to deprecate getArtifacts which is
    returning a byte[] in favor of a new method getArtifactsFiles which
returns a set of file
  * Refactor the internal components to use the new APIs and in
    particular refactor completely the logic of DefaultPresentationBuilder