[Misc] Fixes StaticListClassFieldTest#testItemsEditor for the current integration tests resolution
Reordering items with large label when configuring a static list can be difficult when the item labels are long and the screen width small.
This was making StaticListClassFieldTest#testItemsEditor fail.
We are shortening the reordered item label to make the test pass, until the items UI component is fixed (see XWIKI-18343).