XWIKI-16544: Improve getdocuments template
This change brings major changes to how `getdocuments` works.

Previously, `getdocuments` was permitting to retrieve non-viewable
documents reference whereas this update does not include these results.
A drawback of this change is that a pagination relying on `getdocuments`
can not use offset number and returnedrows as previously. Since the
number of undisclosed items is unknown, the total number of checked rows
is included in the answer in order to be used as the next offset in the
pagination.

Also, totalrows can not be used reliably since it does not differentiate
between viewable and non-viewable documents because this verification is
done after the query.