XWIKI-19125: Introduce a displayedAuthor in XWikiDocument
This commit introduces the main architectural changes for having a
displayedAuthor in XWikiDocument. It actually introduces a new
DocumentAuthors concept to hold the various authors that we want to
provide in XWikiDocument. The idea is also to move towards using
UserReference in XWikiDocument, then the new APIs are returning
UserReference. In order to preserve backward compatibility, this commit
also introduces a resolver component to resolve UserReference to
DocumentReference: this resolver is created as deprecated since it aims
at only being used for deprecated APIs.