XWIKI-19774: Temporary attachments aren't deleted when the session is destroyed
This solution now relies on independent SessionCreatedEvent and
SessionDestroyedEvent: those are xwiki events triggered whenever
HttpSession are created / destroyed. The LegacyActionServlet is now an
HttpSessionListener responsible for trigerring those events.
A new SessionDestroyedListener has been created in filesystem-oldcore
module in order to properly call the disposal of the attachments related
to a session whenever a session is destroyed.