XWIKI-19774: Temporary attachments aren't deleted when the session is destroyed
Introduce an HttpSessionManager component in charge of keeping tracks of
the HttpSessions and of properly disposing them whenever the component
is dispose: this ensures that the sessions are properly disposed and
that their possible listeners are called whenever the application is
stopped.
Implements the storage of the TemporaryAttachmentSession directly in the
HttpSession attributes, and modifies this one to be a HttpSessionBindingListener
so that the data are clean whenever the session is disposed.
Clean up other changes.