XWIKI-19591: Local users of subwikis can't recover password or username
This provides two things: the fix of the bug by handling a wiki part in
the URLs for authentication resource, and a rename of the authentication
resource action (forgot becomes retrieveusername and reset becomes
resetpassword).
Note that the latter changes are not backward compatible, but they were
still unstable (since 13.1RC1).