[Misc] Use template parameter in create page URLs in flamingo skin test
We added a dependency to dashboard-ui in an older commit in which leads
to having a template provided. As a consequence we cannot anymore just
rely on /create/ URLs and expect to obtain the editor: our logic is to ask the user for chosing the
template in such scenario. To avoid that, we can just provide an empty
template parameter to simulate that choice.