XWIKI-20338: Serialize form data using FormData when in-place editing
This allows respecting the standard for query serialization when fields
have several values, as is the case for the xaction parameter, and to
do this serialization using a browser-provided feature.
the field name is repeated for each value instead of having comma-separated
values: key=val1,val2 becomes key=val1&key=val2.