XWIKI-17034: Allow to define different grouping strategy for notifications
The purpose of this commit is to start providing the full architecture
for having different strategies for grouping notifications depending on
users' choices and on the kind of output target (e.g. email, alert, rss,
etc).
This commit provides the new roles for defining a grouping strategy and
for loading the strategy component depending on the user and the target.
By default the grouping strategy is the one that was already existing
and based on on the similarity algorithm. Another strategy is already
provided only based on the type of events.
There's still remaining questions to finish this work, like how to deal
with the number of unread events (as it was before handled by
considering the composite events).
There's also work to do to handle the wiki notification settings for
this, and for presenting the different options to the users.
Finally we probably need to rework the templates so that they would work
with the strategies, or they could take into account the strategies for
the display (e.g. a template per strategy).