XWIKI-20844: Implement an accessible pattern for document extra tabs * Fixed a bug related to class introduced in previous commits. * Fixed the lack of URL change when clicking a panel. * Fixed partially the loss of focus on tab activation. * Added a label and a description to the tab list. * Removed additional history entries when changing panel * Finished updating style.
Note: Still losing focus when making the Ajax request. Can't figure out how to get it back. When there's no ajax request (the panel is already loaded), everything can be handled in tabList.js. User position is still at the right place so it's still possible to navigate back easily with Tab + shiftTab but it's a bit annoying.