XWIKI-17034: Allow to define different grouping strategy for notifications
This commit provides an important change in the design in order to fix a
flaw of the previous solution: in previous solution the number of
composite events obtained would never have matched the requested number
of events wanted by the user. The requested number would have limited
the number of individual events only.
So in order to keep same UX for users, I implement here in
DefaultParametrizedNotificationManager#getEvents the possibility to
retrieve as many composite events as defined by the
NotificationParameters#expectedCount.
Then I also provide distinct caches to ensure to be able to cache either
the composite events, or the events, and to distinguish their count.
Finally I deprecate here the NotificationManager interface in favor of
the ParametrizedNotificationManager which is far more simple to use, and
better to maintain on the long run.