XWIKI-20824: Administration user images must have an alternate text (#2167)
* Added an alt text to user avatars.

Note: The change to solve the PR is in getusers.vm, but other files were changed too to solve the same problem. The solution is the same since the structure around those images is the same. Removed the value for the alt in SolR search, because this would cause a repeated text when reading through the document. See JIRA for the reason why this value should be empty.