XWIKI-17034: Allow to define different grouping strategy for notifications
Previous work was about introducing strategies to group events for creating
composite events. This work is about strategies for chosing how many
mails should be sent for each composite event to notify by email.
Historically the strategy was hardcoded and consisted in putting all
composite events in the same email. Here we provide different
strategies: the previous one, but also a strategy allowing to send one
email per composite event, and another one allowing to send a separate
email specifically for mentions.

That work also improves previous API to use a clear UserReference
instead of a String which is vague.

WIP: unit tests are missing, and we need to provide a new property for
configuring the email grouping strategy.