XWIKI-20648: Notifications Watch are set to off when a subset of events are watched
Watch switches used to have two states: ON and OFF. ON being set when
there was an inclusive filter for all event types (no specific event type set)
and for any format (either email, or alert, or both).
OFF being set when there was no filter (since XWIKI-19070) or when there
was an exclusive filter respecting same rules as above.
Note that switching the watch button to on, was leading to the creation
of a filter for all event types and all formats by default.

As a consequence, if there was an inclusive filter set for a specific
event type, the watch switches were never displaying the information.
Which could lead to mistake.

This work performs the following changes:
  1. The switches are ON only if there's an inclusive filter for all
     event types AND all formats (email + alert): the idea here is to be
     consistent with the filters created by the button itself.
  2. If there is at least one filter matching the location, but for a
     subset of events or formats, the switch now displays a middle
     state, to inform the user that there is already a filter for this
     location
  3. When the button is in the middle state and the user switches it,
     then if the filter is contradictory (e.g. it was an exclusive
     filter and the user wants to watch the location) the filter is kept
     is but is disabled
  4. All those new behaviours directly applies at API level, and not
     only on frontend: so using API to watch/unwatch location will have
     an effect on the filters as explained in 3.