XWIKI-16150: Make livetable suggest filters accessible (#2149)
* Fixed some accessibility issues with the bootstrap daterangePicker
** Fixed DOM order.
** Added aria annotations on the input
** Added a tabindex = 0

TODO: Currently the popup is closed when we tab out of the daterangePicker to the popup itself. It shouldn't close as long as the focus is inside their common parent.

* Fixed some of the daterange picker semantics -- no issue returned by Axe-core
* Fixed daterangepicker presentation: overrode bootstrap style to use our theme to remove contrast issues

Note: The dropdown is not accessible but so poorly indicated that axe-core don't find any violation. The full solution would be to replace the daterangepicker bootstrap element with an accessible version.

* Factorized title generation
* Added a label to the select elements of the last filter types
* Fixed accessibility of the date field
* Removed extra style added that's not necessary
* Removed extra attributes that were added and not necessary.

Note: Since the dropdown would need a complete rework to satisfy accessibility standards, a specific description has been added to the date input, so it can be filled by hand by keyboard users. This is not perfect but it's a significant improvement.