XWIKI-20984: Invitation decline link lacks contrast (#2252)
* Updated error message color to follow color theme

Note: It was a hardcoded color, which is against the code style. However we could argue that the color theme should not influence the style of the invitation sent. It's part of the UI here, so it's important to change, and I don't think it'd have nefarious consequences to follow codestyle here.