XWIKI-20904: The generated translation key for standard notification message is too strict
This work provides a new macro in the notifications macro to compute the
description to display for a composite event. The computation is done as
before, based on the number of users concerned by the event. The main
change is that the macro performs several fallbacks to try to always
display something.