XWIKI-21521: DefaultSkin form should be accessible * Changed the definition list term to an HTML label * Added a javascript function to link the label with its field.
Note: since the ID is built with a prefix that's computed in XWikiDocument.java (L3824), I did not use velocity to fill up the 'for' attribute of the label when building the page. Instead, I preferred changing the value afterwards with a jquery selector. This is similar to the solution used to solve XWIKI-21523