XWIKI-17510: The reset password functionality is not affected by password strength rules
Start working on the integration test to validate XWIKI-17510 and more
globally that the register page was working in a close wiki.
When doing so I discovered I actually also needed XWIKI-21444 to be
fixed and in particular to have live validation. So I refactored
registerinline to be able to use livevalidation and to also handle
properly captcha.

Still a WIP, currently RegisterIT#registerWikiSyntaxName is failing,
still need to investigate why as it seems to work when testing manually.