XWIKI-17510: The reset password functionality is not affected by password strength rules (#2488)
This work mainly provides new Java APIs to access information from RegistrationConfiguration. It also provides a new API to check that the policy for the passwords is respected, and use that new API for user creation. This fixes a bug when creating a new user in private wiki. Note that this PR doesn't provide an API for all information exposed in RegistrationConfiguration: specifically all configuration related to velocity scripts are not exposed.
This PR also improves the default template used for registration to enable live validation, and allows to see captcha, and fixes a few bug when using registration without a RegistrationConfig xobject.
Most of those scenario are now covered in the integration tests which is checking if the registration works in open/close wiki, with and without a RegistrationConfig page.

  * Provide new APIs for accessing RegistrationConfiguration
  * Define a script service to access this API from scripts
  * Rewrite velocity macro #loadConfig to exploit this new API and to
    remove unused arguments
  * Use the new API also in oldcore where needed
  * Provide new API in ResetPasswordManager to allow ensuring a password
    is compliant with the registration rules on backend
  * Fix bugs and perform improvments on templates
  * Add a new integration test for checking password policy
  * Get rid of the testing without livevalidation as we don't have that
    option anymore
  * Improve integration test to also test with and without the
    RegistrationConfig page
---------

Co-authored-by: Manuel Leduc <manuel.leduc@xwiki.com>