[Misc] Fix coverage threshold in modules
Running jacoco with Java 17 leads to a slightly different computation of
the coverage on Enum, probably because of a new implementation of Enum
vs the one in Java 11. As a result we need to fix the threshold in few
modules.

(cherry picked from commit e65865fb899f0d297a75a6be72443348d812822e)