XWIKI-21608: XWIKI-21509 Broke Upgrade if Tomcat Service is Running, But Not Enabled
If the service is running, we need to restart it in any case, no matter if
it's enabled or not.
However is-active does not always seem to work if the package just gets
installed (see XWIKI-18624), so also check if it's only enabled and try to
restart it then.

(cherry picked from commit 6d56ea3ecc97cdd12e96577c669ea08878e6c527)