XWIKI-18971: Number property input does not take into account the current locale to parse floating numbers
Use a NumberFormat instance based on the locale of the owner document
for parsing number values.
I also simplified a bit the code of NumberClass#fromString, started
refactoring of NumberClassTest, but it's still missing lots of test
cases.

Also it's still a WIP: right now we've several unit tests failing in
oldcore because of this: probably because the owner document is now
mandatory for this method. We might need a fallback in case no owner
document is specified.