XWIKI-21779: Admin section: make the search suggest section pass webstandard tests * Updated the template to avoid incorrect checkbox in a label.
There was the error `ERROR: Any “select” descendant of a “label” element with a “for” attribute must have an ID value that matches that “for” attribute. at line [1044] column [105]` which failed for the `activated` property. The previous fix did not solve that so instead I decided to put the checkbox input before the label so that it doesn't fail if it's not properly linked.