XWIKI-20616: Update jstree files and remove reference to throbber.gif
`jstree` less files included in `xwiki-platform-tree-webjar` were coming from version `3.3.5` (or earlier) whereas the version declared in `pom.xml` is `3.3.16`.
This commit resynchronize files originated from `jstree` to match the ones provided in version `3.3.16`.

This commit also comment out the reference to `throbber.gif` in `jstree` `mixins.less` file as the gif file is no part of `xwiki-platform-tree-webjar` and rule is override in `tree.less` to use `spinner.gif`.