XWIKI-21779: Admin section: make the search suggest section pass webstandard tests (#2816)
* XWIKI-21779: Admin section: make the search suggest section pass webstandard tests
* Updated the template to avoid duplicating IDs

Somehow adding the empty object as a parameter changes the ID (adds a {} at the start, still a correct ID, and different from the one used in Config), but does not change how the template is used to create new sources. New sources are still correctly generated.

* XWIKI-21779: Admin section: make the search suggest section pass webstandard tests
* Updated the template to avoid incorrect checkbox in a label.

There was the error `ERROR: Any “select” descendant of a “label” element with a “for” attribute must have an ID value that matches that “for” attribute. at line [1044] column [105]` which failed for the `activated` property. The previous fix did not solve that so instead I decided to put the checkbox input before the label so that it doesn't fail if it's not properly linked.

* Removed the `for` attribute for the labels in the template, because they didn't point towards any input.
* Added an empty dl after checkboxes, in order to keep a correct dd structure.
* Removed the comment from the test for searchSuggest

Note: The best solution would be to replace completely this dd with a regular list, but this is way faster to patch it.

* XWIKI-21779: Admin section: make the search suggest section pass webstandard tests
* Added a comment with a reference to the discussion on the PR

as far as I know, the PRs are kept in storage by GitHub as long as the repository is up, so adding such a link shouldn't pose problem for long term maintenance :)