XWIKI-21452: Macros info, success, warning and error are only distinguished by colors * Fixed faling test in admin tests
Note: It's because in some condition, the registerinline macro is used instead of the register_macros. This registerinline macro relies on the #info macro instead of the {{info}} macro. The first of those two was not updated and now for some reason adds a newline after the false HTML block. I don't know how to register inline, so I couldn't reproduce this use case, and decided to just ignore this line break instead of a space.