XWIKI-22182: Page history is missing after a page containing attachments is moved
Revert "XWIKI-9046: Renaming a document holding a lot of revisions can lead to an OutOfMemory exception (#2926)"

This reverts commit 8e509b70136b2c64359c49ad64ce128874a9c4c0.

The solution provided in this commit wasn't taking into account usecases
when the doc wasn't saved in DB yet, e.g. like in case of renaming. Also
the solution is not good since we actually never set the archive in doc
in case of criteria not all inclusive, and if it had worked it would
have been a problem since we'd have kept only latest revision and lost
the remaining of the history when saving docs. All in all we need
another solution.