XWIKI-17966: Mention events should display the name of the wiki where they took place in Notifications * Removed the second HTML template * Replaced it with jsoup queries.
Note: Did not add the WikiManagerScriptServiceComponentList to the loaded components. It did not change the way the template is rendered AFAICS, and the only use would have been to check that the returned value is the one we expect. For now, we just test that the mention notification properly calls the wiki script service.