XWIKI-21988: Page naming strategy is blocking creation of pages not respecting it
The goal here is to immediately display a feedback in the create page
dialog telling the user that the reference is wrong before they click on
create.

I performed following changes here:
  * Refactor CreateActionRequestHandler to have a clean separation
    between retrieving the reference and performing checks over
templates
  * Refactor a bit CreateAction to perform immediate checks over the
    rights and the name strategy before even checking if the template is
chosen: the goal being here to immediately provide exception if there's
a problem
  * Refactor createinline.vm to inject bootstrap modal html elements
    when used with an ajax request: the idea here is to display an ajax
modal instead of the old deprecated xwiki dialog
  * Refactor xwiki.js to use a jquery ajax call to request and display
    createinline
  * Add listener on xwiki:dom:updated events in both create.js and
    select.js to properly load them in case of dom updates

Some work is still needed to display a better information message to the
user, in particular about the strategy that should be used.