[Misc] Wait more thoroughly for the tree to be loaded
This is an attempt to reduce flickering in tests that might be caused by
proceeding too quickly with loading other tree nodes while the initial
tree nodes are still loading. Then possibly the selection of the root
node might not be cleared correctly before selecting the desired node as
the root node is automatically selected during the initial load.